+++ spec/ductwork/branch/advance_converge_spec.rb
@@ -31,11 +31,15 @@ RSpec.describe Ductwork::Branch, "#advance" do
       pipeline: pipeline
     )
   end
+  let(:transition) { create(:transition, branch:) }
+  let(:advancement) { create(:advancement, transition:) }
   let(:other_branch) { create(:branch, :completed, pipeline:) }
   let(:output_payload) { { payload: }.to_json }
   let(:payload) { 1 }
 
   before do
+    transition
+    advancement
     create(:process, :current)
     other_step = create(
       :step,
@@ -51,7 +55,7 @@ RSpec.describe Ductwork::Branch, "#advance" do
 
   it "creates a new step and enqueues a job" do
     expect do
-      branch.advance!
+      branch.advance!(transition, advancement)
     end.to change(Ductwork::Step, :count).by(1)
       .and change(Ductwork::Job, :count).by(1)
 
@@ -65,13 +69,13 @@ RSpec.describe Ductwork::Branch, "#advance" do
   it "passes the output payloads as input arguments to the next step" do
     allow(Ductwork::Job).to receive(:enqueue)
 
-    branch.advance!
+    branch.advance!(transition, advancement)
 
     expect(Ductwork::Job).to have_received(:enqueue).with(anything, 1)
   end
 
   it "completes the transition and advancement records" do
-    branch.advance!
+    branch.advance!(transition, advancement)
 
     be_almost_now = be_within(1.second).of(Time.current)
     transition = branch.transitions.sole